Add integration snapshot tests for UI rendering (#230)
Add insta-based snapshot tests that render the full UI via ratatui's
TestBackend and compare against committed .snap files. This catches
layout regressions, missing overlays, and rendering bugs automatically.

Tests cover: sidebar layout, chat messages, normal/insert mode indicator,
empty state, help overlay, and narrow terminal auto-hide.

Move populate_demo_data() from main.rs to App method in app.rs with a
base_date parameter for deterministic timestamps in tests.

closes #230
