+++ Cargo.toml
@@ -28,4 +28,5 @@ open = "5"
 notify-rust = "4"
 
 [dev-dependencies]
+insta = "1"
 rstest = "0.25"
